check for local host being in drone list

+++ simics/bin/clean-drones.sh
@@ -11,6 +11,10 @@ aflout=$AFL_DATA/output/$base
 seeds=$AFL_DATA/seeds/$base
 echo "aflout is $aflseed"
 for f in $flist; do
+    if [ $f = $HOSTNAME ]; then
+        echo "Um, your host is in the drones file?, skipping"
+        continue
+    fi
     echo $f
     ssh $USER@$f rm -fr $aflout
     ssh $USER@$f rm -fr $seeds
